src/aws-lambda/generate-products-fanout/requirements.txt
boto3>=1.28.57
pyaml